requirements.txt
PyExecJS==1.5.1
requests==2.24.0
pandas==1.0.5